+++ requests/templates/requests/sent_donation_list.html
@@ -1,28 +1,17 @@
 <!DOCTYPE html>
+{% load static %}
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Dearbebe</title>
-    <link rel="stylesheet" href="/requests/static/requests/common.css" />
-    <link rel="stylesheet" href="/requests/static/requests/received_list.css" />
-    <script defer src="/requests/static/requests/js/received_list.js"></script>
+    <link rel="stylesheet" href="{% static 'requests/common.css' %}" />
+    <link rel="stylesheet" href="{% static 'requests/sent_list' %}" />
+    <script defer src="{% static 'requests/js/sent_list' %}"></script>
   </head>
   <body>
+    <!-- 신청 취소, 삭제도 모달 염두에 두기 -->
     <div class="main-container">
-      <!-- 신청 거절 모달 -->
-      <div class="reject-modal hidden">
-        <p>거절하시겠습니까?</p>
-        <div class="x-divide"></div>
-        <form method="POST" action="">
-          <button type="button" id="cancel">취소</button>
-          <div class="y-divide"></div>
-          <button type="submit" id="ok">확인</button>
-          <!-- 거절 시 받은 신청란에서 사라짐 -->
-        </form>
-      </div>
-      <div class="darken hidden"></div>
-
       <div class="container">
         <!-- 헤더 -->
         <div class="header">
@@ -31,9 +20,10 @@
         <!-- 받은 신청 보낸 신청-->
         <div class="mini-header">
           <div class="request-radio">
-            <a href="#" class="marked">받은 신청</a>
-            <a href="./sent_donation_list.html">보낸 신청</a>
+            <a href="./received_donation_list.html">받은 신청</a>
+            <a href="#" class="marked">보낸 신청</a>
           </div>
+          <!-- 나눔 교환 드롭다운 -->
           <div class="mini-dropdown">
             <div class="selected">
               <p>나눔</p>
@@ -44,14 +34,14 @@
             </div>
             <div class="options hidden">
               <a href="#" class="item">나눔</a>
-              <a href="./received_exchange_list.html" class="item">교환</a>
+              <a href="./sent_exchange_list.html" class="item">교환</a>
             </div>
           </div>
         </div>
         <!-- 날짜에 따른 신청들 구분 -->
         <div class="requests-container">
           <div class="date">6/30</div>
-          <!-- 각 신청 -->
+          <!-- 각 신청(대기중) -->
           <div class="r-container">
             <div class="top-wrapper">
               <div class="text-container">
@@ -95,13 +85,25 @@
               간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
               간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
             </div>
-            <!-- 쪽지 버튼, 거절 버튼 -->
             <div class="button-wrapper">
-              <a href="" class="send-message">쪽지하기</a>
-              <button class="reject">거절</button>
+              <!-- 버튼 - 대기중(신청 취소 버튼) -->
+              <button class="txt-btn">신청 취소</button>
+              <p class="waiting">대기중</p>
+
+              <!-- 거절됨일 때(삭제 버튼) -->
+              <!-- <button class="txt-btn">삭제</button>
+              <p class="rejected">거절됨</p> -->
+
+              <!-- 거래중일 때(버튼 없음) -->
+              <!-- <div></div>
+              <p class="trading">거래중</p> -->
+
+              <!-- 거래 완료일 때(삭제 버튼)-->
+              <!-- <button class="txt-btn">삭제</button>
+              <p class="completed">거래 완료</p> -->
             </div>
           </div>
-          <!-- 각 신청 -->
+          <!-- 각 신청(거절됨) -->
           <div class="r-container">
             <div class="top-wrapper">
               <div class="text-container">
@@ -145,10 +147,61 @@
               간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
               간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
             </div>
-            <!-- 쪽지 버튼, 거절 버튼 -->
+            <!-- 버튼 - 거절됨(삭제 버튼) -->
             <div class="button-wrapper">
-              <a href="" class="send-message">쪽지하기</a>
-              <button class="reject">거절</button>
+              <button class="txt-btn">삭제</button>
+              <p class="rejected">거절됨</p>
+            </div>
+          </div>
+          <!-- 각 신청(거래 중) -->
+          <div class="r-container">
+            <div class="top-wrapper">
+              <div class="text-container">
+                <div class="profile-bar">
+                  <img
+                    src="/requests/static/requests/images/default_profile.png"
+                    alt="기본 프로필 사진"
+                  />
+                  <!-- 프로필 사진이 있다면 <img ... / >-->
+                  <div class="profile">
+                    <!-- 닉네임, 시간, 위치 -->
+                    <div class="nickname-time">
+                      <div class="nickname">닉네임</div>
+                      <div class="time">13:20</div>
+                    </div>
+                    <div class="region">
+                      <img
+                        src="/requests/static/requests/images/place.svg"
+                        alt="위치 아이콘"
+                      />
+                      <p>상계동</p>
+                    </div>
+                  </div>
+                </div>
+                <div class="title">미끄럼 방지 여아 신발</div>
+                <!-- 거래 희망 장소 -->
+                <div class="place">
+                  <div class="dot"></div>
+                  <p>상계동 하나로 마트 앞</p>
+                </div>
+              </div>
+              <!-- 썸네일 -->
+              <div class="image-container">
+                <img
+                  src="/requests/static/requests/images/sample-img.png"
+                  alt="상품 이미지"
+                />
+              </div>
+            </div>
+            <div class="description">
+              간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
+              간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
+            </div>
+            <!-- 버튼 - 거래 중(버튼 없음) -->
+            <div class="button-wrapper">
+              <!-- 형태 깨트리지 않기 위한 div -->
+              <div></div>
+              <p class="trading">거래중</p>
             </div>
           </div>
         </div>
@@ -156,7 +209,7 @@
         <!-- 날짜에 따른 신청들 구분 -->
         <div class="requests-container">
           <div class="date">7/1</div>
-          <!-- 각 신청 -->
+          <!-- 각 신청 (거래 완료) -->
           <div class="r-container">
             <div class="top-wrapper">
               <div class="text-container">
@@ -200,10 +253,10 @@
               간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
               간단하게 입력된 설명 간단하게 입력된 설명 간단하게 입력된 설명
             </div>
-            <!-- 쪽지 버튼, 거절 버튼 -->
+            <!-- 버튼 - 거래 완료(삭제 버튼) -->
             <div class="button-wrapper">
-              <a href="" class="send-message">쪽지하기</a>
-              <button class="reject">거절</button>
+              <button class="txt-btn">삭제</button>
+              <p class="completed">거래 완료</p>
             </div>
           </div>
         </div>
@@ -213,25 +266,16 @@
           <nav>
             <ul>
               <li>
-                <img src="/requests/static/requests/images/home.svg" alt="" />홈
+                <img src="{% static 'requests/images/home.svg' %}" alt="" />홈
               </li>
               <li>
-                <img
-                  src="/requests/static/requests/images/message.svg"
-                  alt=""
-                />쪽지
+                <img src="{% static 'requests/images/message.svg' %}" alt="" />쪽지
+              </li>
+              <li>
+                <img src="{% static 'requests/images/form-active.svg' %}" alt="" />신청 관리
               </li>
               <li class="active">
-                <img
-                  src="/requests/static/requests/images/form-active.svg"
-                  alt=""
-                />신청 관리
-              </li>
-              <li>
-                <img
-                  src="/requests/static/requests/images/my.svg"
-                  alt=""
-                />마이페이지
+                <img src="{% static 'requests/images/my.svg' %}" alt="" />마이페이지
               </li>
             </ul>
           </nav>
