+++ requests/views.py
@@ -6,8 +6,6 @@ from posts.models import Item
 from requests.enums import Status
 from django.utils.timezone import localtime
 from django.contrib.auth.decorators import login_required
-from collections import defaultdict
-from reviews.models import Review
 
 
 # ------------------------
@@ -140,79 +138,3 @@ def manage_sent_exchange_request(request, request_id, action):
 
     return redirect('sent_exchange_requests')
 
-# ------------------------
-# 마이페이지 - 교환/나눔 내역 조회 뷰
-# ------------------------
-
-@login_required
-def my_exchange_history(request):
-    member = request.user
-
-    # 받은 교환: 내가 올린 글에 대해 누군가 신청한 교환 성사된 것
-    received = ExchangeRequest.objects.filter(
-        item__member=member,
-        status=Status.COMPLETED 
-    ).select_related('item', 'member')
-
-    # 보낸 교환: 내가 신청해서 성사된 교환
-    sent = ExchangeRequest.objects.filter(
-        member=member,
-        status=Status.COMPLETED
-    ).select_related('item', 'item__member')
-
-    # 리뷰 작성 여부 설정
-    for req in list(received) + list(sent):
-        req.review_written = Review.objects.filter(exchange_request=req, writer=member).exists()
-
-    # 날짜별 그룹핑
-    all_requests = list(received) + list(sent)
-    all_requests.sort(key=lambda r: r.updated_at, reverse=True)
-
-    grouped = defaultdict(list)
-    for req in all_requests:
-        date = localtime(req.updated_at).strftime("%Y.%m.%d")
-        grouped[date].append(req)
-
-    return render(request, 'requests/my_exchange_history.html', {
-        'grouped': grouped
-    })
-
-@login_required
-def my_sent_donations(request):
-    member = request.user
-
-    completed_requests = DonationRequest.objects.filter(
-        item__member=member,
-        status=Status.COMPLETED
-    ).select_related('item', 'member')
-
-    grouped = defaultdict(list)
-    for req in completed_requests:
-        date = localtime(req.updated_at).strftime("%Y.%m.%d")
-        grouped[date].append(req)
-
-    return render(request, 'requests/my_sent_donations.html', {
-        'grouped': grouped
-    })
-
-
-@login_required
-def my_received_donations(request):
-    member = request.user
-
-    completed_requests = DonationRequest.objects.filter(
-        member=member,
-        status=Status.COMPLETED
-    ).select_related('item', 'item__member')
-
-    for req in completed_requests:
-        req.review_written = Review.objects.filter(donation_request=req, writer=member).exists()
-
-    grouped = defaultdict(list)
-    for req in completed_requests:
-        date = localtime(req.updated_at).strftime("%Y.%m.%d")
-        grouped[date].append(req)
-
-    return render(request, 'requests/my_received_donations.html', {
-        'grouped': grouped
-    })
