+++ reviews/templates/reviews/myreview.html
@@ -1,10 +1,11 @@
 <!DOCTYPE html>
+{% load static %}
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/reviews/static/reviews/common.css" />
-    <link rel="stylesheet" href="/reviews/static/reviews/myreview.css" />
+    <link rel="stylesheet" href="{% static 'reviews/common.css' %}" />
+    <link rel="stylesheet" href="{% static 'reviews/myreview.css' %}" />
 
     <title>Dearbebe</title>
   </head>
@@ -14,10 +15,7 @@
       <!-- 헤더 -->
       <header class="header">
         <button class="back-btn">
-          <img
-            src="/reviews/static/reviews/images/back-btn.svg"
-            alt="뒤로가기"
-          />
+          <img src="{% static 'reviews/images/back-btn.svg' %}" alt="뒤로가기" />
         </button>
         <h1 class="page-title">나의 후기</h1>
       </header>
@@ -63,11 +61,7 @@
           <h3>거래 후기</h3>
           <button class="dropdown-btn">
             작성 후기
-            <img
-              src="/reviews/static/reviews/images/down-icon.svg"
-              alt="화살표"
-              class="down-icon"
-            />
+            <img src="{% static 'reviews/images/down-icon.svg' %}" alt="화살표" class="down-icon" />
           </button>
         </div>
 
@@ -148,11 +142,7 @@
             </div>
             <div class="review-meta">
               <span class="date">3일전</span>
-              <img
-                src="/reviews/static/reviews/images/trash.svg"
-                alt="삭제"
-                class="delete-btn"
-              />
+              <img src="{% static 'reviews/images/trash.svg' %}" alt="삭제" class="delete-btn" />
             </div>
           </li>
         </ul>
@@ -163,23 +153,16 @@
         <nav>
           <ul>
             <li>
-              <img src="/reviews/static/reviews/images/home.svg" alt="" />홈
+              <img src="{% static 'reviews/images/home.svg' %}" alt="" />홈
             </li>
             <li>
-              <img
-                src="/reviews/static/reviews/images/message.svg"
-                alt=""
-              />쪽지
+              <img src="{% static 'reviews/images/message.svg' %}" alt="" />쪽지
             </li>
             <li>
-              <img src="/reviews/static/reviews/images/form.svg" alt="" />신청
-              관리
+              <img src="{% static 'reviews/images/form.svg' %}" alt="" />신청 관리
             </li>
             <li class="active">
-              <img
-                src="/reviews/static/reviews/images/my-active.svg"
-                alt=""
-              />마이페이지
+              <img src="{% static 'reviews/images/my-active.svg' %}" alt="" />마이페이지
             </li>
           </ul>
         </nav>
@@ -198,6 +181,6 @@
       </div>
     </div>
 
-    <script src="/reviews/static/reviews/js/myreview.js"></script>
+    <script src="{% static 'reviews/js/myreview.js' %}"></script>
   </body>
 </html>
