+++ reviews/templates/reviews/sent-share.html
@@ -1,19 +1,20 @@
 <!DOCTYPE html>
+{% load static %}
 <html lang="en">
 
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Dearbebe</title>
-    <link rel="stylesheet" href="/reviews/static/reviews/common.css" />
-    <link rel="stylesheet" href="/reviews/static/reviews/share.css" />
+    <link rel="stylesheet" href="{% static 'reviews/common.css' %}" />
+    <link rel="stylesheet" href="{% static 'reviews/share.css' %}" />
 </head>
 
 <body>
     <div class="main-container">
         <div class="header">
             <button class="back-btn">
-                <img src="/reviews/static/reviews/images/back-btn.svg" alt="">
+                <img src="{% static 'reviews/images/back-btn.svg' %}" alt="" />
             </button>
             <div class="header-title">나눔 내역</div>
         </div>
@@ -23,7 +24,7 @@
         <div class="date-label">6/30</div>
         <button class="dropdown-btn">
                 보낸 나눔
-                <img src="/reviews/static/reviews/images/down-icon.svg" alt="아래 화살표" class="down-icon">
+                <img src="{% static 'reviews/images/down-icon.svg' %}" alt="아래 화살표" class="down-icon" />
             </button></div>
 
         <div class="item-card">
@@ -79,11 +80,19 @@
         <footer class="bottom-nav">
             <nav>
                 <ul>
-                    <li><img src="/reviews/static/reviews/images/home.svg" alt=""><br />홈</li>
-                    <li><img src="/reviews/static/reviews/images/message.svg" alt=""><br />쪽지</li>
-                    <li><img src="/reviews/static/reviews/images/form.svg" alt=""><br />신청 관리</li>
-                    <li><img src="/reviews/static/reviews/images/my.svg" alt=""><br />마이페이지</li>
-                </ul>
+            <li>
+              <img src="{% static 'reviews/images/home.svg' %}" alt="" />홈
+            </li>
+            <li>
+              <img src="{% static 'reviews/images/message.svg' %}" alt="" />쪽지
+            </li>
+            <li>
+              <img src="{% static 'reviews/images/form.svg' %}" alt="" />신청 관리
+            </li>
+            <li class="active">
+              <img src="{% static 'reviews/images/my-active.svg' %}" alt="" />마이페이지
+            </li>
+          </ul>
             </nav>
         </footer>
     </div>
