+++ reviews/views.py
@@ -7,6 +7,12 @@ from .forms import ReviewForm
 from django.db.models import Avg
 from django.utils.timezone import localtime, now
 from requests.models import ExchangeRequest, DonationRequest
+from collections import defaultdict
+from reviews.models import Review
+from requests.enums import Status
+from django.views.decorators.http import require_GET
+from django.http import JsonResponse
+from accounts.models import Member
 
 @login_required
 def create_review(request, request_type, request_id):
@@ -136,3 +142,109 @@ def delete_review(request, review_id):
     review.delete()
     messages.success(request, "리뷰가 삭제되었습니다.")
     return redirect('my_written_reviews')  # 혹은 적절한 리뷰 목록 url name으로 변경
+
+
+
+# --- 마이페이지, 프로필 수정 ---
+
+@login_required
+def my_page(request):
+    return render(request, 'reviews/mypage.html', {'member': request.user})
+
+@login_required
+def edit_profile(request):
+    member = request.user
+    if request.method == 'POST':
+        nickname = request.POST.get('nickname')
+        profile_image = request.FILES.get('profile_image')
+
+        member.nickname = nickname
+        if profile_image:
+            member.profile_image = profile_image
+        member.save()
+        return redirect('my_page')
+
+    return render(request, 'reviews/edit.html', {'member': member})
+
+# ------------------------
+# 마이페이지 - 교환/나눔 내역 조회 뷰
+# ------------------------
+
+@login_required
+def my_exchange_history(request):
+    member = request.user
+
+    # 받은 교환: 내가 올린 글에 대해 누군가 신청한 교환 성사된 것
+    received = ExchangeRequest.objects.filter(
+        item__member=member,
+        status=Status.COMPLETED 
+    ).select_related('item', 'member')
+
+    # 보낸 교환: 내가 신청해서 성사된 교환
+    sent = ExchangeRequest.objects.filter(
+        member=member,
+        status=Status.COMPLETED
+    ).select_related('item', 'item__member')
+
+    # 리뷰 작성 여부 설정
+    for req in list(received) + list(sent):
+        req.review_written = Review.objects.filter(exchange_request=req, writer=member).exists()
+
+    # 날짜별 그룹핑
+    all_requests = list(received) + list(sent)
+    all_requests.sort(key=lambda r: r.updated_at, reverse=True)
+
+    grouped = defaultdict(list)
+    for req in all_requests:
+        date = localtime(req.updated_at).strftime("%Y.%m.%d")
+        grouped[date].append(req)
+
+    return render(request, 'reviews/change.html', {
+        'grouped': grouped
+    })
+
+@login_required
+def my_sent_donations(request):
+    member = request.user
+
+    completed_requests = DonationRequest.objects.filter(
+        item__member=member,
+        status=Status.COMPLETED
+    ).select_related('item', 'member')
+
+    grouped = defaultdict(list)
+    for req in completed_requests:
+        date = localtime(req.updated_at).strftime("%Y.%m.%d")
+        grouped[date].append(req)
+
+    return render(request, 'reviews/sent-share.html', {
+        'grouped': grouped
+    })
+
+
+@login_required
+def my_received_donations(request):
+    member = request.user
+
+    completed_requests = DonationRequest.objects.filter(
+        member=member,
+        status=Status.COMPLETED
+    ).select_related('item', 'item__member')
+
+    for req in completed_requests:
+        req.review_written = Review.objects.filter(donation_request=req, writer=member).exists()
+
+    grouped = defaultdict(list)
+    for req in completed_requests:
+        date = localtime(req.updated_at).strftime("%Y.%m.%d")
+        grouped[date].append(req)
+
+    return render(request, 'reviews/share.html', {
+        'grouped': grouped
+    })
+
+@require_GET
+def check_id_duplicate(request):
+    username = request.GET.get('username')
+    exists = Member.objects.filter(username=username).exists()
+    return JsonResponse({'exists': exists})
